pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "worldreward"
version = "0.1.0"
description = "Physics-verifiable evaluation pipeline for video/3D world models."
readme = "README.md"
requires-python = ">=3.9"
license = { text = "MIT" }
authors = [{ name = "World Reward Contributors" }]
keywords = ["world-models", "evaluation", "physics", "rlvr", "gemini", "veo"]
dependencies = [
  "google-genai>=1.0.0",
  "prompt_toolkit>=3.0.0",
  "pyyaml>=6.0",
  "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
  "mypy>=1.12.0",
  "pytest>=8.0.0",
  "ruff>=0.8.0",
  "types-PyYAML>=6.0.12.20240917",
]

[project.scripts]
worldreward = "worldreward.main:main"

[tool.setuptools]
package-dir = { "" = "src" }
include-package-data = true

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
worldreward = ["builtin_configs/*.yaml"]

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]

[tool.ruff]
target-version = "py39"
line-length = 100

[tool.ruff.lint]
select = ["E4", "E7", "E9", "F", "I"]

[tool.mypy]
python_version = "3.9"
files = ["src", "tests"]
ignore_missing_imports = true
warn_unused_ignores = true
strict_optional = true
